airflow-docker/requirements.txt
apache-airflow==2.9
pyspark==3.3.0
apache-airflow-providers-apache-spark==2.0.0
